+++ CHANGELOG.md
@@ -3,3 +3,11 @@
 ## [1.0.0] - 2023-10-25
 
  - Initial release
+
+## [1.1.0] - 2023-10-30
+
+ - Refactoring added more helper functions
+ - Adding invoice rows with array
+ - Adding mutation rows with array
+ - Throws InputException if errors are returned
+ - Stricter type control